backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
python-multipart
pillow
openai
huggingface_hub
sqlite-utils
pydantic
requests
python-dotenv
# Diğerleri zaten ekliydi
